tushen 2nd

+++ WebTushen/src/main/java/com/tushenshop/model/Category.java
@@ -32,7 +32,7 @@ public class Category {
 
     // Getters and Setters
 
-    public long getCategoryId() {
+    public Integer getCategoryId() {
         return id;
     }
